telegram_bot/requirements.txt
python-telegram-bot[job-queue]>=21.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pytz>=2024.1
